refactor: reconcile direct execution cutover survivors
